Merge branch 'tipc-next'
Jon Maloy says:

====================
tipc: separate link aggregation from link layer

We continue the work on separating the roles of the link aggregation and
link layers, as well as making code cleanups in general.

This second commit batch focuses on moving the orchestration of link
failover and synchronization to the node level, as well as preparing the
node lock structure for further future impovements. We also make some
changes to message delivery between link and socket layer, in order to
make this mechanism safer and less obscure.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 1df33a11454de804661c8e19cd0e464914eefc6d ]