Merge branch 'tcp-loss-probe'
Yuchung Cheng says:

====================
minor tail loss probe improvements

This patch series enhance the tail loss probe (TLP) on some error
conditions. When TLP fails to send a probe, it will no longer
extend the RTO. When it fails to send a new packet because of
receiver window limit, it'll try to retransmit the last packet.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 5bbbcc0073331d4e37c9f91c59cfc9f5d2033860 ]